requirements.txt
alembic==1.7.7
brotlipy==0.7.0
certifi==2021.10.8
# cffi @ file:///C:/ci_310/cffi_1642682485096/work
charset-normalizer==2.0.12
click==8.0.4
cloudpickle==2.0.0
colorama==0.4.4
# cryptography @ file:///C:/ci/cryptography_1639414690059/work
cycler==0.11.0
databricks-cli==0.16.4
docker==5.0.3
entrypoints==0.4
Flask==2.0.3
fonttools==4.30.0
gitdb==4.0.9
GitPython==3.1.27
greenlet==1.1.2
# idna @ file:///tmp/build/80754af9/idna_1637925883363/work
importlib-metadata==4.11.3
importlib-resources==5.4.0
itsdangerous==2.1.1
Jinja2==3.0.3
joblib==1.1.0
kiwisolver==1.4.0
Mako==1.2.0
MarkupSafe==2.1.1
matplotlib==3.5.1
mkl-fft==1.3.1
# mkl-random @ file:///C:/ci/mkl_random_1626186163140/work
mkl-service==2.4.0
mlflow==1.24.0
# numpy @ file:///C:/ci_310/numpy_and_numpy_base_1643798589088/work
packaging==21.3
pandas==1.3.5
Pillow==9.0.1
prometheus-client==0.13.1
prometheus-flask-exporter==0.19.0
protobuf==3.19.4
# pycparser @ file:///tmp/build/80754af9/pycparser_1636541352034/work
# pyOpenSSL @ file:///opt/conda/conda-bld/pyopenssl_1643788558760/work
pyparsing==3.0.7
# PySocks @ file:///C:/ci/pysocks_1594394709107/work
python-dateutil==2.8.2
pytz==2021.3
pywin32==227
PyYAML==6.0
querystring-parser==1.2.4
# requests @ file:///opt/conda/conda-bld/requests_1641824580448/work
scikit-learn==1.0.2
scipy==1.7.3
seaborn==0.11.2
# six @ file:///tmp/build/80754af9/six_1644875935023/work
sklearn==0.0
smmap==5.0.0
SQLAlchemy==1.4.32
sqlparse==0.4.2
# -e git+https://github.com/neerajbafila/pytorch-CNN.git@ed4f8b10a9c73d232acf58baf619837cdea943f3#egg=src
tabulate==0.8.9
threadpoolctl==3.1.0
torch==1.11.0
torchaudio==0.11.0
torchvision==0.12.0
tqdm==4.63.0
typing_extensions==4.1.1
urllib3==1.26.9
waitress==2.1.1
websocket-client==1.3.1
Werkzeug==2.0.3
# win-inet-pton @ file:///C:/ci/win_inet_pton_1605306165655/work
wincertstore==0.2
zipp==3.7.0
